requirements.txt
opencv-python
tensorflow
numpy
matplotlib
pillow
imutils
scikit-learn


